// apis/petstore/tests/types/petstore.ts
export interface User {
  id?: number;
  username: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phone: string;
  userStatus?: number;
}

export interface Category {
  id?: number;
  name: string;
}

export interface Tag {
  id?: number;
  name: string;
}

export interface Pet {
  id?: number;
  name: string;
  category?: Category;
  photoUrls: string[];
  tags?: Tag[];
  status?: 'available' | 'pending' | 'sold';
}

export interface Order {
  id?: number;
  petId: number;
  quantity: number;
  shipDate?: string;
  status?: 'placed' | 'approved' | 'delivered';
  complete?: boolean;
}

export interface ApiResponse {
  code: number;
  type: string;
  message: string;
}